requirements.txt
torch
torchvision
gradio
opencv-python
numpy
Pillow
matplotlib

